+++ libretro-common/formats/json/rjson.c
@@ -982,7 +982,7 @@ size_t rjson_get_source_column(rjson_t *json)
 int rjson_get_source_context_len(rjson_t *json)
 {
    const unsigned char *from = json->input_buf, *to = json->input_end, *p = json->input_p;
-   return ((p + 256 < to ? p + 256 : to) - (p > from + 256 ? p - 256 : from));
+   return (int)(((p + 256 < to ? p + 256 : to) - (p > from + 256 ? p - 256 : from)));
 }
 
 const char* rjson_get_source_context_buf(rjson_t *json)
@@ -1013,7 +1013,7 @@ bool rjson_check_context(rjson_t *json, unsigned int depth, ...)
 
 unsigned int rjson_get_context_depth(rjson_t *json)
 {
-   return json->stack_top - json->stack;
+   return (unsigned int)(json->stack_top - json->stack);
 }
 
 size_t rjson_get_context_count(rjson_t *json)
